Review-agent/requirement.txt
python-dotenv==1.2.1
langgraph>=1.0.0
langchain-core>=1.0.0
langchain>=1.0.0
langchain-openai>=0.3.0
boto3
langchain-aws
graphviz
grandalf
langchain-google-vertexai
langchain-google-genai
langchain-deepseek
langchain-groq
langchain-tavily